estandarizacion de test
